pyproject.toml
[project]
name = "shantabaiagent"
version = "0.1.0"
description = ""
authors = [
    {name = "SuyodhanJ6",email = "prashantmalge181@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = " 3.11.11"
dependencies = [
    "notebook (>=7.3.2,<8.0.0)",
    "langchain (>=0.3.14,<0.4.0)",
    "langchain-groq (>=0.2.3,<0.3.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langgraph-checkpoint (>=2.0.10,<3.0.0)",
    "langgraph (>=0.2.62,<0.3.0)",
    "selenium (>=4.27.1,<5.0.0)",
    "langchain-community (>=0.3.14,<0.4.0)",
    "language-tool-python (>=2.8.1,<3.0.0)",
    "speechrecognition (>=3.14.0,<4.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
